# src/swagger/swagger.yaml
basePath: /
definitions:
...
  api.ContactsResponse:
    properties:
      contacts:
        items:
          $ref: '#/definitions/whatsapp.WhatsappChat'
        type: array
      debug:
        description: Extra interface{} `json:"extra,omitempty"`
        items:
          type: string
        type: array
      status:
        type: string
      success:
        type: boolean
      total:
        type: integer
    type: object
...
  api.InviteResponse:
    properties:
      debug:
        description: Extra interface{} `json:"extra,omitempty"`
        items:
          type: string
        type: array
      status:
        type: string
      success:
        type: boolean
      url:
        type: string
    type: object
  api.IsOnWhatsAppResponse:
    properties:
      debug:
        description: Extra interface{} `json:"extra,omitempty"`
        items:
          type: string
        type: array
      registered:
        items:
          type: string
        type: array
      status:
        type: string
      success:
        type: boolean
      total:
        type: integer
    type: object
...
  api.LIDResponse:
    properties:
      debug:
        description: Extra interface{} `json:"extra,omitempty"`
        items:
          type: string
        type: array
      lid:
        type: string
      phone:
        type: string
      status:
        type: string
      success:
        type: boolean
    type: object
...
  api.ReceiveResponse:
    properties:
      debug:
        description: Extra interface{} `json:"extra,omitempty"`
        items:
          type: string
        type: array
      limit:
        type: integer
      messages:
        items:
          $ref: '#/definitions/whatsapp.WhatsappMessage'
        type: array
      page:
        type: integer
      server:
        $ref: '#/definitions/models.QpServer'
      status:
        type: string
      success:
        type: boolean
      total:
        type: integer
      totalPages:
        type: integer
    type: object
...
  api.SendResponse:
    properties:
      debug:
        description: Extra interface{} `json:"extra,omitempty"`
        items:
          type: string
        type: array
      message:
        $ref: '#/definitions/api.SendResponseMessage'
      status:
        type: string
      success:
        type: boolean
    type: object
...
  api.UserInfoRequest:
    properties:
      jids:
        items:
          type: string
        type: array
    type: object
  api.UserInfoResponse:
    properties:
      debug:
        description: Extra interface{} `json:"extra,omitempty"`
        items:
          type: string
        type: array
      status:
        type: string
      success:
        type: boolean
      total:
        type: integer
      userinfos:
        items: {}
        type: array
    type: object
...
  models.QpResponse:
    properties:
      debug:
        description: Extra interface{} `json:"extra,omitempty"`
        items:
          type: string
        type: array
      status:
        type: string
      success:
        type: boolean
    type: object
...
  whatsapp.WhatsappChat:
    properties:
      id:
        description: (Identifier) whatsapp contact id, based on phone number or timestamp
        type: string
      labels:
        items:
          $ref: '#/definitions/whatsapp.WhatsappChatLabel'
        type: array
      lid:
        description: (Local Identifier) new whatsapp unique contact id
        type: string
      phone:
        description: phone number in E164 format
        type: string
      title:
        type: string
    type: object
...
info:
  contact:
    name: QuePasa Support
    url: https://github.com/inteliagenciadigital/quepasa
  description: QuePasa is a Go-based WhatsApp bot platform that exposes HTTP APIs
    for WhatsApp messaging integration
  license:
    name: GNU Affero General Public License v3.0
    url: https://github.com/inteliagenciadigital/quepasa/blob/main/LICENSE.md
  termsOfService: https://github.com/inteliagenciadigital/quepasa
  title: QuePasa WhatsApp API
  version: 5.26.0625.0
paths:
...
  /command:
    get:
      consumes:
      - application/json
      description: Execute control commands for the bot server (start, stop, restart)
      parameters:
      - description: Command action
        enum:
        - start
        - stop
        - restart
        - direct
        - groups
        - broadcasts
        - readreceipts
        - deliveryreceipts
        - calls
        - debug
        in: query
        name: action
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Execute bot commands
      tags:
      - Bot
...
  /contacts:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all WhatsApp contacts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ContactsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Get contacts
      tags:
      - Contacts
...
  /download:
    get:
      description: Downloads media files (images, videos, documents) from WhatsApp
        messages
      parameters:
      - description: Message ID (query parameter alternate)
        in: query
        name: messageid
        type: string
      - description: Use cached content
        in: query
        name: cache
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Media file
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Download media
      tags:
      - Download
...
  /invite:
    get:
      consumes:
      - application/json
      description: Generates an invite link for a specific WhatsApp group
      parameters:
      - description: Chat ID (query parameter)
        in: query
        name: chatid
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.InviteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Generate group invite link
      tags:
      - Groups
...
  /isonwhatsapp:
    post:
      consumes:
      - application/json
      description: Checks if provided phone numbers are registered on WhatsApp
      parameters:
      - description: Phone numbers to check
        in: body
        name: request
        required: true
        schema:
          properties:
            phones:
              items:
                type: string
              type: array
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.IsOnWhatsAppResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Check WhatsApp registration
      tags:
      - Contacts
...
  /paircode:
    get:
      description: Generates a pairing code for WhatsApp authentication using phone
        number
      parameters:
      - description: Phone number for pairing
        in: query
        name: phone
        required: true
        type: string
      - description: Days of message history to sync
        in: query
        name: historysyncdays
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Generate pairing code
      tags:
      - Connection
...
  /receive:
    get:
      consumes:
      - application/json
      description: Retrieves pending messages from WhatsApp with optional cache filters
      parameters:
      - description: Timestamp filter for messages
        in: query
        name: timestamp
        type: string
      - description: 'Filter by exceptions error status: ''true'' for messages with
          exceptions errors, ''false'' for messages without exceptions errors, omit
          for all messages'
        in: query
        name: exceptions
        type: string
      - description: Filter by message type (supports comma-separated list)
        in: query
        name: type
        type: string
      - description: 'Filter by category: sent, received, sync, unhandled, events'
        in: query
        name: category
        type: string
      - description: Search text in id, chat, text, trackid, participant and exceptions
        in: query
        name: search
        type: string
      - description: 'Filter by fromme boolean: true or false'
        in: query
        name: fromme
        type: string
      - description: 'Filter by fromhistory boolean: true or false'
        in: query
        name: fromhistory
        type: string
      - description: Filter by chat id (contains)
        in: query
        name: chatid
        type: string
      - description: Filter by message id (contains)
        in: query
        name: messageid
        type: string
      - description: Filter by track id (contains)
        in: query
        name: trackid
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReceiveResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Receive messages
      tags:
      - Message
...
  /scan:
    get:
      description: Generates a QR code for WhatsApp Web authentication scanning
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Generate QR code
      tags:
      - Connection
  /send:
    post:
      consumes:
      - application/json
      description: |-
        Endpoint to send messages via WhatsApp. Accepts sending of:
        - Plain text (field "text")
        - Files by URL (field "url") â€” server will download and send as attachment
        - Base64 content (field "content") â€” use format data:<mime>;base64,<data>
        - Polls (field "poll") â€” send the poll JSON in the "poll" field
        - Location (field "location") â€” send location with latitude/longitude in the "location" object
        - Contact (field "contact") â€” send contact with phone/name in the "contact" object
        - Sticker (field "sticker") â€” send a sticker from URL or base64 content; auto-converted to WebP 512Ã—512 via FFmpeg

        Main fields:
        - chatId: chat identifier (can be WID, LID or number with suffix @s.whatsapp.net)
        - text: message text
        - url: public URL to download a file
        - content: embedded base64 content (e.g.: data:image/png;base64,...)
        - fileName: file name (optional, used when name cannot be inferred)
        - poll: JSON object with the poll (question, options, selections)
        - location: JSON object with location data (latitude, longitude, name, address, url)
        - contact: JSON object with contact data (phone, name, vcard)
        - sticker: JSON object with sticker source (url or content as base64/data URI)

        Location object fields:
        - latitude (float64, required): Location latitude in degrees (e.g.: -23.550520)
        - longitude (float64, required): Location longitude in degrees (e.g.: -46.633308)
        - name (string, optional): Location name/description
        - address (string, optional): Location full address
        - url (string, optional): URL with link to the map

        Contact object fields:
        - phone (string, required): Contact phone number
        - name (string, required): Contact display name
        - vcard (string, optional): Full vCard string (auto-generated if not provided)

        Sticker object fields:
        - url (string): Public URL of the sticker image/video to download and convert
        - content (string): Base64-encoded content or data URI (e.g.: data:image/png;base64,...)
        Note: images and videos are automatically converted to WebP 512Ã—512 using FFmpeg.
        Animated formats (video, gif, apng) produce animated WebP stickers (max 10s, 15fps).
        Static images produce static WebP stickers.

        Examples:
        Text:
        ```json
        {
        "chatId": "5511999999999@s.whatsapp.net",
        "text": "Hello, world!"
        }
        ```
        Poll:
        ```json
        {
        "chatId": "5511999999999@s.whatsapp.net",
        "poll": {
        "question": "Which languages do you know?",
        "options": ["JavaScript","Python","Go","Java","C#","Ruby"],
        "selections": 3
        }
        }
        ```
        Location:
        ```json
        {
        "chatId": "5511999999999@s.whatsapp.net",
        "location": {
        "latitude": -23.550520,
        "longitude": -46.633308,
        "name": "Avenida Paulista, SÃ£o Paulo"
        }
        }
        ```
        Contact:
        ```json
        {
        "chatId": "5511999999999@s.whatsapp.net",
        "contact": {
        "phone": "5511999999999",
        "name": "John Doe"
        }
        }
        ```
        Base64:
        ```json
        {
        "chatId": "5511999999999@s.whatsapp.net",
        "content": "data:image/png;base64,...."
        }
        ```
        File by URL:
        ```json
        {
        "chatId": "5511999999999@s.whatsapp.net",
        "url": "https://example.com/path/to/file.jpg"
        }
        ```
        Sticker by URL:
        ```json
        {
        "chatId": "5511999999999@s.whatsapp.net",
        "sticker": {
        "url": "https://example.com/sticker.png"
        }
        }
        ```
        Sticker by base64:
        ```json
        {
        "chatId": "5511999999999@s.whatsapp.net",
        "sticker": {
        "content": "data:image/png;base64,...."
        }
        }
        ```
      parameters:
      - description: Request body. Use 'content' for base64, 'url' for remote files,
          'poll' for poll JSON, 'location' for location object, 'contact' for contact
          object, or 'sticker' for sticker object.
        in: body
        name: request
        schema:
          properties:
            chatId:
              type: string
            contact:
              properties:
                name:
                  type: string
                phone:
                  type: string
                vcard:
                  type: string
              type: object
            content:
              type: string
            fileName:
              type: string
            location:
              properties:
                address:
                  type: string
                latitude:
                  type: number
                longitude:
                  type: number
                name:
                  type: string
                url:
                  type: string
              type: object
            poll:
              properties:
                options:
                  items:
                    type: string
                  type: array
                question:
                  type: string
                selections:
                  type: integer
              type: object
            sticker:
              properties:
                content:
                  type: string
                url:
                  type: string
              type: object
            text:
              type: string
            url:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SendResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.SendResponse'
      security:
      - ApiKeyAuth: []
      summary: Send any type of message (text, file, poll, base64 content, location,
        contact, sticker)
      tags:
      - Send
...
  /spam:
    post:
      consumes:
      - application/json
      description: Send messages using any available server (spam/broadcast functionality)
      parameters:
      - description: Spam message request
        in: body
        name: request
        required: true
        schema:
          properties:
            chatId:
              type: string
            text:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SendResponse'
        "423":
          description: No server available
          schema:
            $ref: '#/definitions/api.SendResponse'
      security:
      - ApiKeyAuth: []
      summary: Send spam messages
      tags:
      - Application
...
  /useridentifier:
    get:
      consumes:
      - application/json
      description: Retrieves the Local Identifier (LID) for a given phone number
      parameters:
      - description: Phone number
        in: query
        name: phone
        type: string
      - description: Local identifier
        in: query
        name: lid
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.LIDResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Get user identifier (LID)
      tags:
      - Contacts
  /userinfo:
    post:
      consumes:
      - application/json
      description: Retrieves detailed information for WhatsApp users by their JIDs
      parameters:
      - description: User info request with JIDs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.UserInfoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.UserInfoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Get user information
      tags:
      - Contacts
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-QUEPASA-TOKEN
    type: apiKey
swagger: "2.0"